Remove PatternPickerCubit
