+++ lib/widgets/editors/arranger/pattern_picker/pattern_picker.dart
@@ -17,8 +17,9 @@
   along with Anthem. If not, see <https://www.gnu.org/licenses/>.
 */
 
+import 'package:anthem/model/project.dart';
+import 'package:anthem/widgets/project/project_controller.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:anthem/theme.dart';
 import 'package:anthem/widgets/basic/button.dart';
@@ -27,142 +28,151 @@ import 'package:anthem/widgets/basic/clip/clip.dart';
 import 'package:anthem/widgets/basic/controls/vertical_scale_control.dart';
 import 'package:anthem/widgets/basic/icon.dart';
 import 'package:anthem/widgets/basic/scroll/scrollbar.dart';
-import 'package:anthem/widgets/editors/arranger/pattern_picker/pattern_picker_cubit.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:provider/provider.dart';
 
-class PatternPicker extends StatelessWidget {
+enum PatternFilterKind { midi, audio, automation }
+
+class PatternPicker extends StatefulWidget {
+  const PatternPicker({Key? key}) : super(key: key);
+
+  @override
+  State<PatternPicker> createState() => _PatternPickerState();
+}
+
+class _PatternPickerState extends State<PatternPicker> {
+  double patternHeight = 50;
+
   final ScrollController scrollController = ScrollController();
-
-  PatternPicker({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<PatternPickerCubit, PatternPickerState>(
-      builder: (context, state) {
-        return NotificationListener<SizeChangedLayoutNotification>(
-          onNotification: (notification) {
-            WidgetsBinding.instance.addPostFrameCallback((_) {
-              scrollController.position.notifyListeners();
-            });
-            return true;
-          },
-          child: SizeChangedLayoutNotifier(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: [
-                SizedBox(
-                  height: 26,
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: [
-                      Expanded(
-                        child: ButtonTabs(
-                          tabs: [
-                            ButtonTabDef.withIcon(
-                              icon: Icons.midi,
-                              id: PatternFilterKind.midi,
-                            ),
-                            ButtonTabDef.withIcon(
-                              icon: Icons.audio,
-                              id: PatternFilterKind.audio,
-                            ),
-                            ButtonTabDef.withIcon(
-                              icon: Icons.automation,
-                              id: PatternFilterKind.automation,
-                            ),
-                          ],
+    final project = Provider.of<ProjectModel>(context);
+    final projectController = Provider.of<ProjectController>(context);
+
+    return NotificationListener<SizeChangedLayoutNotification>(
+      onNotification: (notification) {
+        WidgetsBinding.instance.addPostFrameCallback((_) {
+          scrollController.position.notifyListeners();
+        });
+        return true;
+      },
+      child: SizeChangedLayoutNotifier(
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            SizedBox(
+              height: 26,
+              child: Row(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: [
+                  Expanded(
+                    child: ButtonTabs(
+                      tabs: [
+                        ButtonTabDef.withIcon(
+                          icon: Icons.midi,
+                          id: PatternFilterKind.midi,
                         ),
-                      ),
-                      const SizedBox(width: 4),
-                      VerticalScaleControl(
-                        min: 25,
-                        max: 60,
-                        value: state.patternHeight,
-                        onChange: (value) {
-                          final cubit = context.read<PatternPickerCubit>();
-                          cubit.setPatternHeight(value);
-                          scrollController.position.notifyListeners();
-                        },
-                      ),
-                    ],
+                        ButtonTabDef.withIcon(
+                          icon: Icons.audio,
+                          id: PatternFilterKind.audio,
+                        ),
+                        ButtonTabDef.withIcon(
+                          icon: Icons.automation,
+                          id: PatternFilterKind.automation,
+                        ),
+                      ],
+                    ),
                   ),
-                ),
-                const SizedBox(height: 4),
-                Expanded(
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: [
-                      Expanded(
-                        child: Container(
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(4),
-                            border: Border.all(
-                              color: Theme.panel.border,
-                              width: 1,
-                            ),
-                            color: Theme.panel.accentDark,
-                          ),
-                          child: ClipRRect(
-                            borderRadius: BorderRadius.circular(4),
-                            child: SingleChildScrollView(
-                              controller: scrollController,
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.stretch,
-                                children: state.patternIDs
-                                    .map(
-                                      (patternID) => Padding(
-                                        padding:
-                                            const EdgeInsets.only(bottom: 1),
-                                        child: SizedBox(
-                                          height: state.patternHeight,
-                                          child: Clip.fromPattern(
-                                            patternID: patternID,
-                                            ticksPerPixel: 5,
-                                          ),
+                  const SizedBox(width: 4),
+                  VerticalScaleControl(
+                    min: 25,
+                    max: 60,
+                    value: patternHeight,
+                    onChange: (value) {
+                      setState(() {
+                        patternHeight = value;
+                      });
+                      scrollController.position.notifyListeners();
+                    },
+                  ),
+                ],
+              ),
+            ),
+            const SizedBox(height: 4),
+            Expanded(
+              child: Row(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: [
+                  Expanded(
+                    child: Container(
+                      decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(4),
+                        border: Border.all(
+                          color: Theme.panel.border,
+                          width: 1,
+                        ),
+                        color: Theme.panel.accentDark,
+                      ),
+                      child: ClipRRect(
+                        borderRadius: BorderRadius.circular(4),
+                        child: SingleChildScrollView(
+                          controller: scrollController,
+                          child: Observer(builder: (context) {
+                            return Column(
+                              crossAxisAlignment: CrossAxisAlignment.stretch,
+                              children: project.song.patternOrder
+                                  .map(
+                                    (patternID) => Padding(
+                                      padding: const EdgeInsets.only(bottom: 1),
+                                      child: SizedBox(
+                                        height: patternHeight,
+                                        child: Clip.fromPattern(
+                                          patternID: patternID,
+                                          ticksPerPixel: 5,
                                         ),
                                       ),
-                                    )
-                                    .toList(),
-                              ),
-                            ),
+                                    ),
+                                  )
+                                  .toList(),
+                            );
+                          }),
+                        ),
+                      ),
+                    ),
+                  ),
+                  const SizedBox(width: 4),
+                  SizedBox(
+                    width: 17,
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.stretch,
+                      children: [
+                        Expanded(
+                          child: Scrollbar(
+                            controller: scrollController,
+                            crossAxisSize: 17,
+                            direction: ScrollbarDirection.vertical,
                           ),
                         ),
-                      ),
-                      const SizedBox(width: 4),
-                      SizedBox(
-                        width: 17,
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.stretch,
-                          children: [
-                            Expanded(
-                              child: Scrollbar(
-                                controller: scrollController,
-                                crossAxisSize: 17,
-                                direction: ScrollbarDirection.vertical,
-                              ),
-                            ),
-                            const SizedBox(height: 4),
-                            Button(
-                              startIcon: Icons.add,
-                              height: 17,
-                              variant: ButtonVariant.ghost,
-                              contentPadding: const EdgeInsets.all(0),
-                              onPress: () {
-                                final cubit =
-                                    context.read<PatternPickerCubit>();
-                                cubit.addPattern();
-                              },
-                            ),
-                          ],
+                        const SizedBox(height: 4),
+                        Button(
+                          startIcon: Icons.add,
+                          height: 17,
+                          variant: ButtonVariant.ghost,
+                          contentPadding: const EdgeInsets.all(0),
+                          onPress: () {
+                            projectController.addPattern();
+                          },
                         ),
-                      ),
-                    ],
+                      ],
+                    ),
                   ),
-                ),
-              ],
+                ],
+              ),
             ),
-          ),
-        );
-      },
+          ],
+        ),
+      ),
     );
   }
 }
