--- lib/widgets/editors/arranger/pattern_picker/pattern_picker_state.dart
@@ -1,15 +1,0 @@
-part of 'pattern_picker_cubit.dart';
-
-// Workaround for https://github.com/rrousselGit/freezed/issues/653
-@Freezed(makeCollectionsUnmodifiable: false)
-class PatternPickerState with _$PatternPickerState {
-  factory PatternPickerState({
-    required ID projectID,
-    required List<ID> patternIDs,
-    required Map<ID, String> patternNames,
-    required double patternHeight,
-    @Default(PatternFilterKind.midi) PatternFilterKind filterType,
-  }) = _PatternPickerState;
-}
-
-enum PatternFilterKind { midi, audio, automation }
