+++ lib/widgets/editors/pattern_editor/pattern_editor_controller.dart
@@ -22,41 +22,12 @@ import 'dart:ui';
 import 'package:anthem/commands/pattern_commands.dart';
 import 'package:anthem/commands/project_commands.dart';
 import 'package:anthem/helpers/id.dart';
-import 'package:anthem/model/pattern/pattern.dart';
 import 'package:anthem/model/project.dart';
 
 class PatternEditorController {
   final ProjectModel project;
 
   PatternEditorController({required this.project});
-
-  ID addPattern([String? name]) {
-    if (name == null) {
-      final patterns = project.song.patterns.nonObservableInner;
-      var patternNumber = patterns.length;
-
-      final existingNames = patterns.values.map((pattern) => pattern.name);
-
-      do {
-        patternNumber++;
-        name = "Pattern $patternNumber";
-      } while (existingNames.contains(name));
-    }
-
-    final patternModel = PatternModel.create(name: name, project: project);
-
-    project.execute(
-      AddPatternCommand(
-        project: project,
-        pattern: patternModel,
-        index: project.song.patternOrder.length,
-      ),
-    );
-
-    project.song.setActivePattern(patternModel.id);
-
-    return patternModel.id;
-  }
 
   void addGenerator(String name, Color color) {
     project.execute(AddGeneratorCommand(
